Add Business Classes

+++ src/main/java/com/darigan/lil/learningspring/data/Reservation.java
@@ -12,10 +12,10 @@ public class Reservation {
     private long reservationId;
 
     @Column(name="ROOM_ID")
-    private int roomId;
+    private long roomId;
 
     @Column(name="GUEST_ID")
-    private int guestId;
+    private long guestId;
 
     @Column(name="RES_DATE")
     private Date date;
@@ -28,19 +28,19 @@ public class Reservation {
         this.reservationId = reservationId;
     }
 
-    public int getRoomId() {
+    public long getRoomId() {
         return roomId;
     }
 
-    public void setRoomId(int roomId) {
+    public void setRoomId(long roomId) {
         this.roomId = roomId;
     }
 
-    public int getGuestId() {
+    public long getGuestId() {
         return guestId;
     }
 
-    public void setGuestId(int guestId) {
+    public void setGuestId(long guestId) {
         this.guestId = guestId;
     }
